// src/common/types.ts
export type RunCommand = { command: string; args: any[] };

export interface IPostWorkbenchCreationActions {
    readonly openUri?: UriComponents;
    runCommands?: RunCommand[];
}

export interface IPreWorkbenchDisposalActions {
    readonly runCommands?: RunCommand[];
}

export interface IRouteResult {
    workspace: IWorkspace;
    workbenchOptions?: IWorkbenchConstructionOptions;
    readonly onDidCreateWorkbench?: IPostWorkbenchCreationActions;
    readonly onWillDisposeWorkbench?: IPreWorkbenchDisposalActions;
    welcomeBanner?: IWelcomeBanner;
    suppressDefaultRemoteExtensions?: true;
}

// https://github.com/microsoft/vscode/blob/ae0f691839eff7dc0354224f7cff9ed3a128cf54/src/vscode-dts/vscode.d.ts#L14530
export interface AuthenticationSession {
    readonly id: string;
    readonly accessToken: string;
    readonly account: {
        label: string;
        id: string;
    };
    readonly scopes: readonly string[];
}

export interface MicrosoftAuthenticationSession {
    readonly id: string;
    readonly account: {
        label: string;
        id: string;
    };
    readonly scopes: readonly string[];
    getAccessToken(): Promise<string>;
}

export interface IMicrosoftAuthentication {
    getSessions(scopes: string[], options?: { clientId?: string; forceNewSession?: boolean }): Promise<MicrosoftAuthenticationSession[]>;
}